refactor(shared-memory): modernize core library
- Replace header guards with #pragma once across all headers
- Add const qualifiers to value parameters
- Switch brace initialization to = style for scalar members
- Use [[maybe_unused]] attribute instead of (void) casts
- Replace std::memcpy with std::copy_n in monitor test
- Add TypedDict annotations to format_benchmark_stats.py
- Improve SkipIfBusy and GetLatestBuffer documentation
- Add Mermaid architecture diagrams to README
- Refine BUILD.bazel sanitizer and coverage test exclusions
- Clean up trailing whitespace across source files

Action-Id: A-0001
Evidence: bazel run check -- format PASS, lint PASS, test-prod 6/6 PASS

Assisted by Claude

+++ dex/infrastructure/shared_memory/shared_memory_streaming_test.cc
@@ -928,7 +928,7 @@ TEST_F(SharedMemStreamingTest, LargeBufferCommunication) {
         dex::shared_memory::Producer<LargeBuffer> producer{shared_memory_name};
         producer.Run([](LargeBuffer& buffer, uint /*counter*/, int /*buffer_id*/) {
           buffer.a = 0xDEADBEEF;
-          std::memset(buffer.data.data(), static_cast<int>(0xAA), buffer.data.size());
+          std::memset(buffer.data.data(), 0xAA, buffer.data.size());
         });
         CoverageSafeExit(0);
       },
@@ -950,4 +950,3 @@ TEST_F(SharedMemStreamingTest, LargeBufferCommunication) {
 }
 
 }  // namespace
-
